+++ CHANGELOG.md
@@ -71,6 +71,10 @@ leave();
 
 # v1.4.8
 
+### `create-liveblocks-app`
+
+- Add Comments hooks and options to `--init` command.
+
 ### `@liveblocks/react-comments`
 
 - Improve default styles: